refactor(project): further refactoring and adjustments of docs
