--- crates/project/src/data/internal.rs
@@ -1,2 +1,0 @@
-
-
